+++ source/Makefile.Objects
@@ -212,7 +212,7 @@ OBJS_ESOLVER=esolver.o\
     esolver_ks.o\
     esolver_fp.o\
     esolver_ks_pw.o\
-    esolver_ks_lip.o\
+    esolver_ks_lcaopw.o\
     esolver_sdft_pw.o\
     esolver_lj.o\
     esolver_dp.o\
