+++ source/module_esolver/esolver_ks_lcaopw.cpp
@@ -1,4 +1,4 @@
-#include "esolver_ks_lip.h"
+#include "esolver_ks_lcaopw.h"
 
 #include "module_hamilt_pw/hamilt_pwdft/elecond.h"
 #include "module_io/input_conv.h"
